restructured dash atomic implementations, types and algorithms
