+++ dash/include/dash/GlobRef.h
@@ -3,8 +3,6 @@
 
 #include <dash/GlobMem.h>
 #include <dash/Init.h>
-#include <dash/algorithm/Operation.h>
-#include <dash/Atomic.h>
 
 namespace dash {
 
@@ -46,7 +44,7 @@ public:
 private:
   typedef GlobRef<T>
     self_t;
-  
+
 public:
   /**
    * Default constructor, creates an GlobRef object referencing an element in
@@ -55,7 +53,7 @@ public:
   GlobRef()
   : _gptr(DART_GPTR_NULL) {
   }
-  
+
   /**
    * Constructor, creates an GlobRef object referencing an element in global
    * memory.
@@ -66,7 +64,7 @@ public:
     GlobPtr<T, PatternT> & gptr)
   : GlobRef(gptr.dart_gptr())
   { }
-  
+
   /**
    * Constructor, creates an GlobRef object referencing an element in global
    * memory.
@@ -77,7 +75,7 @@ public:
     const GlobPtr<T, PatternT> & gptr)
   : GlobRef(gptr.dart_gptr())
   { }
-  
+
   /**
    * Constructor, creates an GlobRef object referencing an element in global
    * memory.
@@ -87,7 +85,7 @@ public:
   {
     DASH_LOG_TRACE_VAR("GlobRef(dart_gptr_t)", dart_gptr);
   }
-  
+
   /**
    * Copy constructor.
    */
@@ -96,7 +94,7 @@ public:
     const GlobRef<T> & other)
   : _gptr(other._gptr)
   { }
-  
+
   /**
    * Assignment operator.
    */
@@ -115,33 +113,33 @@ public:
   //  _gptr = other._gptr;
   //  return *this;
   }
-  
+
   inline bool operator==(const self_t & other) const noexcept
   {
     return _gptr == other._gptr;
   }
-  
+
   inline bool operator!=(const self_t & other) const noexcept
   {
     return !(*this == other);
   }
-  
+
   inline bool operator==(const T & value) const
   {
     return static_cast<T>(*this) == value;
   }
-  
+
   inline bool operator!=(const T & value) const
   {
     return !(*this == value);
   }
-  
+
   friend void swap(GlobRef<T> a, GlobRef<T> b) {
     T temp = (T)a;
     a = b;
     b = temp;
   }
-  
+
   T get() const {
     DASH_LOG_TRACE("T GlobRef.get()", "explicit get");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
@@ -150,7 +148,7 @@ public:
     dart_get_blocking(static_cast<void *>(&t), _gptr, ds.nelem, ds.dtype);
     return t;
   }
-  
+
   operator T() const {
     DASH_LOG_TRACE("GlobRef.T()", "conversion operator");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
@@ -159,41 +157,41 @@ public:
     dart_get_blocking(static_cast<void *>(&t), _gptr, ds.nelem, ds.dtype);
     return t;
   }
-  
+
   void get(T *tptr) const {
     DASH_LOG_TRACE("GlobRef.get(T*)", "explicit get into provided ptr");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
     dart_storage_t ds = dash::dart_storage<T>(1);
     dart_get_blocking(static_cast<void *>(tptr), _gptr, ds.nelem, ds.dtype);
   }
-  
+
   void get(T& tref) const {
     DASH_LOG_TRACE("GlobRef.get(T&)", "explicit get into provided ref");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
     dart_storage_t ds = dash::dart_storage<T>(1);
     dart_get_blocking(static_cast<void *>(&tref), _gptr, ds.nelem, ds.dtype);
   }
-  
+
   void put(T& tref) const {
     DASH_LOG_TRACE("GlobRef.put(T&)", "explicit put of provided ref");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
     dart_storage_t ds = dash::dart_storage<T>(1);
     dart_put_blocking(_gptr, static_cast<void *>(&tref), ds.nelem, ds.dtype);
   }
-  
+
   void put(T* tptr) const {
     DASH_LOG_TRACE("GlobRef.put(T*)", "explicit put of provided ptr");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
     dart_storage_t ds = dash::dart_storage<T>(1);
     dart_put_blocking(_gptr, static_cast<void *>(tptr), ds.nelem, ds.dtype);
   }
-  
+
   operator GlobPtr<T>() const {
     DASH_LOG_TRACE("GlobRef.GlobPtr()", "conversion operator");
     DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
     return GlobPtr<T>(_gptr);
   }
-  
+
   GlobRef<T> & operator=(const T val) {
     DASH_LOG_TRACE_VAR("GlobRef.=()", val);
     DASH_LOG_TRACE_VAR("GlobRef.=", _gptr);
@@ -203,7 +201,7 @@ public:
     dart_put_blocking(_gptr, static_cast<const void *>(&val), ds.nelem, ds.dtype);
     return *this;
   }
-  
+
   GlobRef<T> & operator+=(const T& ref)
   {
   #if 0
@@ -224,21 +222,21 @@ public:
   #endif
     return *this;
   }
-  
+
   GlobRef<T> & operator-=(const T& ref) {
     T val  = operator T();
     val   -= ref;
     operator=(val);
     return *this;
   }
-  
+
   GlobRef<T> & operator++() {
     T val = operator T();
     ++val;
     operator=(val);
     return *this;
   }
-  
+
   GlobRef<T> operator++(int) {
     GlobRef<T> result = *this;
     T val = operator T();
@@ -246,14 +244,14 @@ public:
     operator=(val);
     return result;
   }
-  
+
   GlobRef<T> & operator--() {
     T val = operator T();
     --val;
     operator=(val);
     return *this;
   }
-  
+
   GlobRef<T> operator--(int) {
     GlobRef<T> result = *this;
     T val = operator T();
@@ -261,28 +259,28 @@ public:
     operator=(val);
     return result;
   }
-  
+
   GlobRef<T> & operator*=(const T& ref) {
     T val  = operator T();
     val   *= ref;
     operator=(val);
     return *this;
   }
-  
+
   GlobRef<T> & operator/=(const T& ref) {
     T val  = operator T();
     val   /= ref;
     operator=(val);
     return *this;
   }
-  
+
   GlobRef<T> & operator^=(const T& ref) {
     T val  = operator T();
     val   ^= ref;
     operator=(val);
     return *this;
   }
-  
+
   #if 0
   // Might lead to unintended behaviour
   GlobPtr<T> operator &() {
@@ -292,7 +290,7 @@ public:
   dart_gptr_t dart_gptr() const {
     return _gptr;
   }
-  
+
   #if 0
   template<
     typename X=T,
@@ -305,7 +303,7 @@ public:
     return val[pos];
   }
   #endif
-  
+
   /**
    * Checks whether the globally referenced element is in
    * the calling unit's local memory.
@@ -313,7 +311,7 @@ public:
   bool is_local() const {
     return GlobPtr<T>(_gptr).is_local();
   }
-  
+
   /**
    * Get a global ref to a member of a certain type at the
    * specified offset
@@ -327,7 +325,7 @@ public:
     GlobPtr<MEMTYPE> gptr(dartptr);
     return GlobRef<MEMTYPE>(gptr);
   }
-  
+
   /**
    * Get the member via pointer to member
    */
@@ -338,253 +336,18 @@ public:
     size_t offs = (size_t) &( reinterpret_cast<P*>(0)->*mem);
     return member<MEMTYPE>(offs);
   }
-  
+
 private:
 
   dart_gptr_t _gptr;
 
 };
-
-
-/**
- * Specialization for atomic values. All atomic operations are 
- * \c const as the \c GlobRef does not own the atomic values.
- */
-template<typename T>
-class GlobRef<dash::Atomic<T>>
-{
-  template<typename U>
-  friend std::ostream & operator<<(
-    std::ostream & os,
-    const GlobRef<U> & gref);
-  
-public:
-  typedef T value_type;
-  
-private:
-  typedef dash::Atomic<T>      atomic_t;
-  typedef GlobRef<atomic_t>      self_t;
-  
-public:
-  /**
-   * Default constructor, creates an GlobRef object referencing an element in
-   * global memory.
-   */
-  GlobRef()
-  : _gptr(DART_GPTR_NULL) {
-  }
-  
-  /**
-   * Constructor, creates an GlobRef object referencing an element in global
-   * memory.
-   */
-  template<typename PatternT>
-  GlobRef(
-    /// Pointer to referenced object in global memory
-    GlobPtr<atomic_t, PatternT> & gptr)
-  : GlobRef(gptr.dart_gptr())
-  { }
-  
-  /**
-   * Constructor, creates an GlobRef object referencing an element in global
-   * memory.
-   */
-  template<typename PatternT>
-  GlobRef(
-    /// Pointer to referenced object in global memory
-    const GlobPtr<atomic_t, PatternT> & gptr)
-  : GlobRef(gptr.dart_gptr())
-  { }
-  
-  /**
-   * Constructor, creates an GlobRef object referencing an element in global
-   * memory.
-   */
-  explicit GlobRef(dart_gptr_t dart_gptr)
-  : _gptr(dart_gptr)
-  {
-    DASH_LOG_TRACE_VAR("GlobRef(dart_gptr_t)", dart_gptr);
-  }
-  
-  /**
-   * Copy constructor.
-   */
-  GlobRef(
-    /// GlobRef instance to copy.
-    const GlobRef<atomic_t> & other)
-  : _gptr(other._gptr)
-  { }
-  
-  /**
-   * Assignment operator.
-   */
-  self_t & operator=(const self_t & other)
-  {
-    DASH_LOG_TRACE_VAR("GlobRef.=()", other);
-    // This results in a dart_put, required for STL algorithms like
-    // std::copy to work on global ranges.
-    // TODO: Not well-defined:
-    //       This violates copy semantics, as
-    //         GlobRef(const GlobRef & other)
-    //       copies the GlobRef instance while
-    //         GlobRef=(const GlobRef & other)
-    //       puts the value.
-    return *this = static_cast<atomic_t>(other);
-  //  _gptr = other._gptr;
-  //  return *this;
-  }
-  
-  inline bool operator==(const self_t & other) const noexcept
-  {
-    return _gptr == other._gptr;
-  }
-  
-  inline bool operator!=(const self_t & other) const noexcept
-  {
-    return !(*this == other);
-  }
-  
-  inline bool operator==(const T & value) const
-  {
-    return (load() == value);
-  }
-  
-  inline bool operator!=(const T & value) const
-  {
-    return !(*this == value);
-  }
-  
-  operator T() const {
-    return load();
-  }
-  
-  operator GlobPtr<T>() const {
-    DASH_LOG_TRACE("GlobRef.GlobPtr()", "conversion operator");
-    DASH_LOG_TRACE_VAR("GlobRef.T()", _gptr);
-    return GlobPtr<atomic_t>(_gptr);
-  }
-  
-  dart_gptr_t dart_gptr() const {
-    return _gptr;
-  }
-  
-  /**
-   * Checks whether the globally referenced element is in
-   * the calling unit's local memory.
-   */
-  bool is_local() const {
-    return GlobPtr<T>(_gptr).is_local();
-  }
-  
-  /// atomically assigns value
-  GlobRef<atomic_t> operator=(const T & value) const {
-    dash::atomic::store(*this, value);
-    return *this;
-  }
-  
-  void store(const T & value) const {
-    dash::atomic::store(*this, value);
-  }
-  /// atomically fetches value
-  T load() const {
-    return dash::atomic::load(*this);
-  }
-  
-  /**
-   * atomically exchanges value
-   */
-  T exchange(const T & value) const {
-    return dash::atomic::exchange(*this, value);
-  }
-  
-  /**
-   * Atomically compares the value with the value of expected and if those are
-   * bitwise-equal, replaces the former with desired.
-   * 
-   * \note not implemented yet
-   * 
-   * \return  True if value is exchanged
-   * 
-   * \see \c dash::atomic::compare_exchange
-   */
-  bool compare_exchange(const T & expected, const T & desired) const {
-    return dash::atomic::compare_exchange(*this, expected, desired);
-  }
-  
-  /*
-   * ---------------------------------------------------------------------------
-   * ------------ specializations for atomic integral types --------------------
-   * ---------------------------------------------------------------------------
-   *
-   *  As the check for integral type is already implemented in the 
-   *  dash::atomic::* algorithms, no check is performed here
-   */
-  
-  /**
-   * DASH specific variant which is faster than \cfetch_add
-   * but does not return value
-   */
-  void add(const T & value) const {
-    dash::atomic::add(*this, value);
-  }
-  
-  T fetch_add(const T & value) const {
-    return dash::atomic::fetch_add(*this, value);
-  }
-  
-  /**
-   * DASH specific variant which is faster than \cfetch_sub
-   * but does not return value
-   */
-  void sub(const T & value) const {
-    dash::atomic::sub(*this, value);
-  }
-  
-  T fetch_sub(const T & value) const {
-    return dash::atomic::fetch_sub(*this, value);
-  }
-  
-  /// prefix atomically increment value by one
-  T operator++ () const {
-    return fetch_add(1) + 1;
-  }
-  
-  /// postfix atomically increment value by one
-  T operator++ (int) const {
-    return fetch_add(1);
-  }
-  
-  /// prefix atomically decrement value by one
-  T operator-- () const {
-    return fetch_sub(1) - 1;
-  }
-  
-  /// postfix atomically decrement value by one
-  T operator-- (int) const {
-    return fetch_sub(1);
-  }
-  
-  /// atomically increment value by ref
-  T operator+=(const T & value) const {
-    return fetch_add(value) + value;
-  }
-  
-  /// atomically decrement value by ref
-  T operator-=(const T & value) const {
-    return fetch_sub(value) - value;
-  }
-  
-private:
-  
-  dart_gptr_t _gptr;
-  
-};
-
 
 template<typename T>
 std::ostream & operator<<(
-std::ostream & os,
-const GlobRef<T> & gref) {
+  std::ostream     & os,
+  const GlobRef<T> & gref)
+{
   char buf[100];
   sprintf(buf,
   "(%08X|%04X|%04X|%016lX)",
